requirement.txt
ultralytics
streamlit
streamlit-webrtc
opencv-python-headless
av
pillow
numpy